requirements.txt
gradio
PyPDF2
pinecone-client
sentence-transformers
cohere

